requirements.txt
# Core Discord Dependencies
discord.py[voice]==2.3.2  # Added voice extras for voice features
PyNaCl==1.5.0  # Required for voice support
python-dotenv==1.0.0

# Database
aiosqlite==0.19.0
SQLAlchemy==2.0.25  # Added for better database management
alembic==1.13.1  # Added for database migrations

# Image Processing
Pillow==10.2.0  # For welcome images and ticket transcripts

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Networking
aiohttp==3.9.1
async-timeout==4.0.3
websockets==12.0  # Added for better WebSocket handling
requests==2.31.0  # Added for health checks

# Utilities
attrs==23.1.0
charset-normalizer==3.3.2
idna==3.4
multidict==6.0.4
yarl==1.9.4

# Logging and Monitoring
colorlog==6.8.0  # Added for colored console output
structlog==24.1.0  # Added for structured logging
psutil==5.9.8  # Added for system metrics
prometheus-client==0.19.0  # Added for metrics export
statsd==4.0.1  # Added for StatsD metrics
datadog==0.48.0  # Added for Datadog integration

# Performance
uvloop==0.19.0; platform_system != "Windows"  # Added for better async performance on Unix
orjson==3.9.10  # Added for faster JSON handling
ujson==5.9.0  # Added for faster JSON handling in health checks
cachetools==5.3.2  # Added for in-memory caching

# Security
cryptography==41.0.7  # Added for enhanced security features
bcrypt==4.1.2  # Added for password hashing
PyJWT==2.8.0  # Added for JWT handling
argon2-cffi==23.1.0  # Added for password hashing

# Development Tools
black==23.12.1  # Added for code formatting
isort==5.13.2  # Added for import sorting
mypy==1.8.0  # Added for type checking
pytest==7.4.4  # Added for testing
pytest-asyncio==0.23.3  # Added for async tests
pytest-cov==4.1.0  # Added for test coverage
pytest-mock==3.12.0  # Added for mocking in tests
freezegun==1.4.0  # Added for time-based testing

# Documentation
Sphinx==7.2.6  # Added for documentation generation
sphinx-rtd-theme==2.0.0  # Added for documentation theme
sphinx-autodoc-typehints==1.25.2  # Added for type hints in docs

# Deployment
docker==7.0.0  # Added for Docker integration
docker-compose==1.29.2  # Added for Docker Compose integration
gunicorn==21.2.0  # Added for production WSGI server
supervisor==4.2.5  # Added for process management
